requirements.txt
anthropic>=0.50.0
streamlit>=1.35.0
python-dotenv>=1.0.0
gspread>=6.0.0
google-auth>=2.0.0
